[feat] footer 추가
- 기존에 만들어놓은 Footer branch merge 후 Footer 추가
